basic ckms client for boostrap server
