+++ Cargo.toml
@@ -44,7 +44,12 @@ redis = { version = "0.23", features = [
     "connection-manager",
     "tokio-comp",
 ] }
-reqwest = { version = "0.11", features = ["json", "native-tls"] }
+reqwest = { version = "0.11", features = [
+    "json",
+    "multipart",
+    "native-tls",
+    "stream",
+] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.96"
 strum = "0.25"
@@ -54,8 +59,8 @@ time = "0.3"
 tiny-keccak = "2.0"
 # It is important to keep this version in sync with that of actix-web
 # use `cargo tree -i tokio` to check there is no other version of tokio
-
 tokio = { version = "1.25", features = ["full"] }
+tokio-util = { version = "0.7", features = ["codec"] }
 tracing = "0.1"
 tracing-log = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
