+++ crate/client/Cargo.toml
@@ -13,4 +13,6 @@ josekit = { workspace = true }
 reqwest = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
 thiserror = { workspace = true }
